requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pandas>=2.1.1
numpy>=1.26.0
scipy>=1.10.1
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
duckdb>=0.8.1
openai>=1.3.0
python-dotenv>=1.0.0
pillow>=10.0.0
kaleido>=0.2.1
openpyxl>=3.1.2
xlrd>=2.0.1
PyPDF2>=3.0.0
PyMuPDF>=1.23.5
python-multipart>=0.0.6
werkzeug>=2.3.7
